+++ doc/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 Fixed
 ^^^^^
 - Error and critical check results directly raise an exception when ``raise_exceptions=True``.
+- Use ``ComplexAttribute`` instead of ``MultiValuedComplexAttribute`` to generate data.
 
 [0.2.2] - 2025-08-13
 --------------------